Working version, completed a lot more api routes
