+++ docs/api.md
@@ -3,7 +3,8 @@
 All Get results are returned as JSON.
 
 
-```/api/users```
+
+## ```/api/users```
 
 - ```GET /all``` - Get all users.
 Returns the whole users table.
@@ -17,10 +18,15 @@ Returns all data (*) about the user.
 - ```GET /:userID``` - Get the user, given a user ID.
 Returns all data (*) about the user.
 
-- Link get Edits, Maps, Fav here
+- ```GET /:userID/maps``` - Get all the maps a user owns.
+
+- ```GET /:userID/favs``` - Get all the maps a user favorites.
+
+- //TODO repercise this ```GET /:userID/edits``` - Get the edit history of the user.
 
 
-```/api/maps```
+
+## ```/api/maps```
 
 - ```GET /all``` - Get all maps.
 Returns all data (*), with the full name of the creator, and the number of favorites.
@@ -34,17 +40,18 @@ Returns all data (*), with the full name of the creator
 - ```GET /:mapID``` - Get the map, given a map ID.
 Returns all data (*) about the map.
 
+- //TODO Implement ```GET /:mapID/points```
+
 - ```DELETE /:mapID``` - Delete the map from database, given a map ID.
 
-- Link get points here
 
 
-```/api/points```
+## ```/api/points```
 
-- Get ID
+- //TODO Get data by ID
 
 
-```/api/favs```
+## ```/api/favs```
 
 - ```POST /u/:userID/m/:mapID``` - Create a new "favorite" relationship, with given userID and mapID.
 
@@ -52,16 +59,17 @@ Returns all data (*) about the map.
 
 - ```DELETE /:favID``` - Delete the relationship from database, given a fav ID.
 
-- Get ID
+- //TODO Get data by ID
 
 
-```/api/edits```
+
+## ```/api/edits```
 
 - ```GET /u/:userID``` - Get all maps edited by the user, given a user ID.
 
-- Get ID
+- //TODO Get data by ID
 
-- Create new edit.
+- //TODO Create new edit.
